Refactor token generation and password utilities by removing unused code and ensuring proper exports
