+++ server/src/utils/hashPassword.ts
@@ -1,6 +1,14 @@
-const validateEmail = (email: string): boolean => {
-  const re = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-  return re.test(email.toLowerCase());
+import bcrypt from 'bcrypt';
+
+
+export const hashPassword = async (plain: string): Promise<string> => {
+  const salt = await bcrypt.genSalt(10);
+  return await bcrypt.hash(plain, salt);
 };
 
-export default validateEmail;
+export const comparePasswords = async (
+  input: string,
+  hashed: string
+): Promise<boolean> => {
+  return await bcrypt.compare(input, hashed);
+};
